% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_connect_function.R
\name{hdx_connect}
\alias{hdx_connect}
\title{Connect to HDX}
\usage{
hdx_connect()
}
\description{
This function uses ckanr to connect to the humanitarian data
exchange.
}
\examples{
hdx_connect()
}
