% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_connect_test_function.R
\name{hdx_connect_test}
\alias{hdx_connect_test}
\title{Connect to HDX}
\usage{
hdx_connect_test(...)
}
\description{
This function uses ckanr to connect to the test version of humanitarian data
exchange.
}
\examples{
hdx_connect_test()
}
