% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_dataset_create_function.R
\name{hdx_dataset_create}
\alias{hdx_dataset_create}
\title{Create a Dataset on HDX}
\usage{
hdx_dataset_create(name = NULL, title = NULL, author = NULL,
  author_email = NULL, maintainer = NULL, maintainer_email = NULL,
  license_id = NULL, notes = NULL, package_url = NULL, version = NULL,
  state = "active", type = NULL, resources = NULL, tags = NULL,
  extras = NULL, relationships_as_object = NULL,
  relationships_as_subject = NULL, groups = NULL, owner_org = NULL,
  key = ckanr::get_default_key(), url = ckanr::get_default_url(),
  dataset_source = NULL, as = "list", ...)
}
\description{
This function will create a dataset on HDX. The original function was created by ROpenSci in their CKANR package.
HDX version simply has an additional argument for dataset_source.
It requires a name, a dataset_source, an owner_org.
}
\examples{
hdx_dataset_create()
}
