% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_dataset_search_function.R
\name{hdx_dataset_search}
\alias{hdx_dataset_search}
\title{HDX Dataset Search Function}
\usage{
hdx_dataset_search(term = "*:*", rows = 10, start = NULL, exact = FALSE)
}
\arguments{
\item{term}{Title to search for. No default}

\item{rows}{Max number of rows to return. Default is 10}

\item{exact}{If exact = TRUE, will return only those datasets that exactly match the term provided. Default is FALSE}
}
\description{
This function will create a data frame of search results from HDX
}
\examples{
hdx_dataset_search()
}
\keyword{hdx}
\keyword{search}
