% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_package_search_function.R
\name{hdx_package_search}
\alias{hdx_package_search}
\title{HDX Package Search Function}
\usage{
hdx_package_search(term = "*:*", rows = 10, start = NULL)
}
\arguments{
\item{term}{Title to search for. No default}

\item{rows}{Max number of rows to return. Default is 10}
}
\description{
This function will create a data frame of search results from HDX
}
\examples{
hdx_package_search()
}
\keyword{hdx}
\keyword{search}
