% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_resource_csv.R
\name{hdx_resource_csv}
\alias{hdx_resource_csv}
\title{HDX Resource CSV Function}
\usage{
hdx_resource_csv(resources)
}
\arguments{
\item{resources}{Results of hdx_resource_list}
}
\description{
This function modify the results of a hdx_resource_list function. It will download csvs when available, It will return results in a nested dataframe.
}
\examples{
hdx_resource_list()
}
\keyword{hdx}
\keyword{resources}
