% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_resource_shapefile_function.R
\name{hdx_resource_shapefile}
\alias{hdx_resource_shapefile}
\title{Download and load shapefiles as simple features}
\usage{
hdx_resource_shapefile(resources)
}
\arguments{
\item{resources}{Results of hdx_resource_list}
}
\description{
Humanitarian data exchange stores it's shapefiles as zipped files This function takes the results of the hdx_dataset_search,
after being expanded by hdx_resource_list function, and downloads the first shapefile to the working directory.
It will create a new folder with the name of the shapefile, and unzip the shapefiles into it.
It will then load the first shapefile from this directory into the environment, in simple features format.
}
\examples{
dataset <- hdx_dataset_search()
resources <- hdx_resource_list(dataset)
shapefile <- hdx_resource_shapefile(resources)
}
