% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_connect_function.R
\name{hdx_connect}
\alias{hdx_connect}
\title{HDX Connect Function}
\usage{
hdx_connect()
}
\description{
This function connects you to the CKAN server for HDX
}
\examples{
hdx_connect()
}
\keyword{connect}
\keyword{hdx}
