% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_dataset_list_function.R
\name{hdx_dataset_list}
\alias{hdx_dataset_list}
\title{HDX Dataset List Function}
\usage{
hdx_dataset_list(limit = 5000)
}
\arguments{
\item{limit}{Limit of packages to search for. Defaults to 5000}
}
\description{
This function will create a data frame of package titles
}
\examples{
hdx_dataset_list()
}
\keyword{hdx}
\keyword{list}
