% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_dataset_search_function.R
\name{hdx_dataset_search}
\alias{hdx_dataset_search}
\title{Search HDX for Datasets}
\usage{
hdx_dataset_search(term = "*:*", rows = 10, start = NULL, exact = FALSE)
}
\arguments{
\item{term}{Title to search for. No default}

\item{rows}{Max number of rows to return. Default is 10}

\item{exact}{If exact = TRUE, will return only those datasets that exactly match the term provided. Default is FALSE}
}
\description{
This function will search the humanitarian data exchange for the term provided to it. It will return a dataframe of the results.
If you are using the precise term and do not want any similar results, use the exact argument. If you want greater than ten results,
use the rows argument.
}
\examples{
hdx_dataset_search()
hdx_dataset_search("afghanistan-roads", exact = TRUE)
hdx_dataset_search(rows = 100)
}
