% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_resource_csv_function.R
\name{hdx_resource_csv}
\alias{hdx_resource_csv}
\title{Download and load CSV Resources}
\usage{
hdx_resource_csv(resources)
}
\arguments{
\item{resources}{Results of hdx_resource_list}
}
\description{
This function modify the results of a hdx_resource_list function.
It will download csvs when available, and return them in a nested dataframe.
}
\examples{
hdx_dataset_search() \%>\% hdx_resource_list() \%>\% hdx_resource_csv()
}
