% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_resource_list_function.R
\name{hdx_resource_list}
\alias{hdx_resource_list}
\title{Expand resources from HDX Dataset Search}
\usage{
hdx_resource_list(package)
}
\arguments{
\item{package}{Results of hdx_dataset_search in dataframe format}
}
\description{
This function modify the results of a hdx_dataset_search function.
It will expand the resources section and left join these results onto the original dataframe.
}
\examples{
dataset <- hdx_dataset_search()
resources <- hdx_resource_list(dataset)
}
