% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_resource_shapefile_function.R
\name{hdx_resource_shapefile}
\alias{hdx_resource_shapefile}
\title{HDX Resource Shapefile Function}
\usage{
hdx_resource_shapefile(resources)
}
\arguments{
\item{resources}{Results of hdx_resource_list}
}
\description{
This function modify the results of a hdx_resource_list function. It will download the first zipped shapefile found, and load it as a simple features dataframe.
}
\examples{
hdx_resource_shapefile()
}
\keyword{hdx}
\keyword{resources}
